+++ changelog.md
@@ -82,6 +82,7 @@
 - ``parseOct`` and ``parseBin`` in parseutils now also support the ``maxLen`` argument similar to ``parseHexInt``
 - Added the proc ``flush`` for memory mapped files.
 - Added the ``MemMapFileStream``.
+- Added ``countup`` and ``countdown`` iterators for floats.
 
 ### Library changes